move getConnectionTimeout from DataSourceFactory to a new JdbcConnector class
