+++ airbyte-cdk/java/airbyte-cdk/core/src/test/java/io/airbyte/cdk/db/factory/DSLContextFactoryTest.java
@@ -7,6 +7,7 @@ package io.airbyte.cdk.db.factory;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import io.airbyte.cdk.integrations.JdbcConnector;
 import java.util.Map;
 import javax.sql.DataSource;
 import org.jooq.DSLContext;
@@ -52,7 +53,7 @@ class DSLContextFactoryTest extends CommonFactoryTest {
         container.getJdbcUrl(),
         dialect,
         connectionProperties,
-        DataSourceFactory.CONNECT_TIMEOUT_DEFAULT);
+        JdbcConnector.CONNECT_TIMEOUT_DEFAULT);
     assertNotNull(dslContext);
     assertEquals(dialect, dslContext.configuration().dialect());
   }
